Implement GameState.get_legal_moves()
This is currently untested!

+++ src/freecell/position.rs
@@ -3,7 +3,7 @@ use std::fmt::{Display, Formatter, Result};
 
 
 pub enum Position {
-    Cascade(u8),
+    Cascade(usize),
     Foundations,
     Freecells
 }
